turbine: put solvers in same file
