+++ steady/test_cases/turbine_array/run_adapt.py
@@ -8,7 +8,7 @@ import os
 import sys
 
 from adapt_utils.steady.test_cases.turbine_array.options import TurbineArrayOptions
-from adapt_utils.steady.swe.turbine.solver import AdaptiveSteadyTurbineProblem
+from adapt_utils.swe.turbine.solver import AdaptiveSteadyTurbineProblem
 from adapt_utils.plotting import *
