+++ steady/test_cases/turbine_array/run_uniform_convergence.py
@@ -4,7 +4,7 @@ import argparse
 import h5py
 import os
 
-from adapt_utils.steady.swe.turbine.solver import AdaptiveSteadyTurbineProblem
+from adapt_utils.swe.turbine.solver import AdaptiveSteadyTurbineProblem
 from adapt_utils.steady.test_cases.turbine_array.options import TurbineArrayOptions
